build(travis): update deploy doc step
only deploy docs if job was triggered from master branch

+++ tools/gh-pages-publish.ts
@@ -1,31 +1,38 @@
-const { cd, exec, echo, touch } = require("shelljs")
-const { readFileSync } = require("fs")
-const url = require("url")
+const { cd, exec, echo, touch } = require('shelljs')
+const { readFileSync } = require('fs')
+const url = require('url')
 
-let repoUrl
-let pkg = JSON.parse(readFileSync("package.json") as any)
-if (typeof pkg.repository === "object") {
-  if (!pkg.repository.hasOwnProperty("url")) {
-    throw new Error("URL does not exist in repository section")
+const branchName = process.env.TRAVIS_BRANCH
+echo(`running on branch ${branchName}`)
+
+if (branchName === 'master') {
+  const pkg = JSON.parse(readFileSync('package.json') as any)
+  let repoUrl
+  if (typeof pkg.repository === 'object') {
+    if (!pkg.repository.hasOwnProperty('url')) {
+      throw new Error('URL does not exist in repository section')
+    }
+    repoUrl = pkg.repository.url
+  } else {
+    repoUrl = pkg.repository
   }
-  repoUrl = pkg.repository.url
+
+  const parsedUrl = url.parse(repoUrl)
+  const repository = (parsedUrl.host || '') + (parsedUrl.path || '')
+  const ghToken = process.env.GH_TOKEN
+
+  echo('Deploying docs!!!')
+  cd('dist/docs')
+  touch('.nojekyll')
+  exec('git init')
+  exec('git add .')
+  exec('git config user.name "Michael Wittwer"')
+  exec('git config user.email "michael.wittwer@shiftcode.ch"')
+  exec('git commit -m "docs(docs): update gh-pages"')
+  exec(
+    `git push --force --quiet "https://${ghToken}@${repository}" master:gh-pages`
+  )
+  echo('Docs deployed!!')
 } else {
-  repoUrl = pkg.repository
+  echo('Not running on master, therefor docs are not deployed')
 }
-
-let parsedUrl = url.parse(repoUrl)
-let repository = (parsedUrl.host || "") + (parsedUrl.path || "")
-let ghToken = process.env.GH_TOKEN
-
-echo("Deploying docs!!!")
-cd("dist/docs")
-touch(".nojekyll")
-exec("git init")
-exec("git add .")
-exec('git config user.name "Michael Wittwer"')
-exec('git config user.email "michael.wittwer@shiftcode.ch"')
-exec('git commit -m "docs(docs): update gh-pages"')
-exec(
-  `git push --force --quiet "https://${ghToken}@${repository}" master:gh-pages`
-)
-echo("Docs deployed!!")
